+++ Game/level_3.h
@@ -161,7 +161,7 @@ private:
     int counterTriangle=0;
     int highscore = 0;
     int counterTogether = 0;
-    int leveltime;
+    double leveltime;
     bool win = false;
     int counterTriangleangle1=0;
     int counterTriangleangle2=0;
